pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "coral"
dynamic = ["version"]
description = "Orchestration system for autonomous coding agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11, <3.14"
authors = [
    { name = "CORAL Team" }
]
keywords = [
    "ai",
    "agents",
    "autonomous",
    "orchestration",
    "evaluation",
    "llm"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "httpx>=0.27",
    "uvicorn>=0.30.0",
    "fastapi>=0.100.0",
    "litellm[proxy]==1.82.3",
    "pip>=26.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]
ui = [
    "starlette>=0.40.0",
    "uvicorn[standard]>=0.30.0",
    "pyyaml>=6.0",
]
all = [
    "coral[dev,ui]",
]

[project.scripts]
coral = "coral.cli:main"

[project.urls]
Homepage = "https://github.com/Human-Agent-Society/CORAL"
Repository = "https://github.com/Human-Agent-Society/CORAL"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "coral/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["coral"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
